# swagger.yaml
openapi: 3.0.0
info:
  title: Saint-Pierre API
  version: 1.0.0
  description: API complète pour la gestion des devoirs, cours et messagerie scolaire
  contact:
    name: Support
  x-logo:
    url: https://your-api-logo.png
servers:
  - url: http://localhost:3000
    description: Serveur de développement
  - url: https://api.saint-pierre.com
    description: Serveur de production
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from /auth/sign-in
  schemas:
...
    Matiere:
      type: object
      properties:
        id_matiere:
          type: string
          pattern: '^[0-9]+$'
          example: '1'
        nom_matiere:
          type: string
          maxLength: 50
          example: 'Mathématiques'
        description:
          type: string
          nullable: true
        couleur:
          type: string
          format: color
          example: '#FF5733'
          nullable: true
        icon_url:
          type: string
          nullable: true
        devoir_icon_url:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
...
    Cours:
      type: object
      properties:
        id_cours:
          type: string
          pattern: '^[0-9]+$'
        nom_cours:
          type: string
          maxLength: 50
        description_cours:
          type: string
          nullable: true
        id_user:
          type: string
          pattern: '^[0-9]+$'
        id_matiere:
          type: string
          pattern: '^[0-9]+$'
          nullable: true
        id_classe:
          type: string
          pattern: '^[0-9]+$'
          nullable: true
        id_specialite:
          type: string
          pattern: '^[0-9]+$'
          nullable: true
        id_option:
          type: string
          pattern: '^[0-9]+$'
          nullable: true
        matiere:
          $ref: '#/components/schemas/Matiere'
        ressources:
          type: array
          items:
            $ref: '#/components/schemas/RessourceCours'
    RessourceCours:
      type: object
      properties:
        id_ressource:
          type: string
          pattern: '^[0-9]+$'
        nom_fichier:
          type: string
        chemin_fichier:
          type: string
        type_fichier:
          type: string
        taille_octets:
          type: string
          pattern: '^[0-9]+$'
        date_ajout:
          type: string
          format: date-time

    # ========== DEVOIR ==========
...
    Error400:
      type: object
      properties:
        error:
          type: string
          example: 'Paramètres invalides'
    Error401:
      type: object
      properties:
        error:
          type: string
          example: 'Accès refusé: token manquant'
    Error403:
      type: object
      properties:
        error:
          type: string
          example: 'Accès refusé: rôle insuffisant'
    Error404:
      type: object
      properties:
        error:
          type: string
          example: 'Ressource non trouvée'
...
    SuccessResponse:
      type: object
      properties:
        message:
          type: string
          example: 'Opération réussie'
...
security:
  - BearerAuth: []
paths:
...
  /api/cours:
    get:
      summary: Lister les cours (filtrables par matière, classe, spécialité ou option)
      tags: [Cours]
      operationId: getCours
      parameters:
        - name: kind
          in: query
          description: Type de filtre (specialite, option)
          schema:
            type: string
        - name: id
          in: query
          description: ID pour le filtre
          schema:
            type: string
        - name: id_matiere
          in: query
          description: ID de la matière
          schema:
            type: string
        - name: id_specialite
          in: query
          description: ID de la spécialité
          schema:
            type: string
        - name: id_option
          in: query
          description: ID de l'option
          schema:
            type: string
      responses:
        '200':
          description: Liste des cours
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Cours'
        '401':
          description: Non authentifié
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error401'

    post:
      summary: Créer un nouveau cours (Professeur uniquement)
      tags: [Cours]
      operationId: createCours
      x-roles: [professeur]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                nom_cours:
                  type: string
                  maxLength: 50
                description_cours:
                  type: string
                id_matiere:
                  type: string
                  pattern: '^[0-9]+$'
                id_classe:
                  type: string
                  pattern: '^[0-9]+$'
                  nullable: true
                id_specialite:
                  type: string
                  pattern: '^[0-9]+$'
                  nullable: true
                id_option:
                  type: string
                  pattern: '^[0-9]+$'
                  nullable: true
                fichiers:
                  type: array
                  items:
                    type: string
                    format: binary
              required: [nom_cours, id_matiere]
              oneOf:
                - required: [id_matiere]
                - required: [id_specialite]
                - required: [id_option]
      responses:
        '201':
          description: Cours créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cours'
        '400':
          description: Données invalides
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400'
  /api/cours/{id}:
    get:
      summary: Récupérer un cours par ID
      tags: [Cours]
      operationId: getCoursById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
      responses:
        '200':
          description: Cours trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cours'
        '404':
          description: Cours non trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404'

    delete:
      summary: Supprimer un cours (Professeur propriétaire)
      tags: [Cours]
      operationId: deleteCours
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
      responses:
        '200':
          description: Cours supprimé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          description: Cours non trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404'
        '403':
          description: Accès refusé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error403'
  /api/cours/matiere/{matiereId}:
    get:
      summary: Lister les cours d'une matière
      tags: [Cours]
      operationId: getCoursByMatiere
      parameters:
        - name: matiereId
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
      responses:
        '200':
          description: Liste des cours
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Cours'
  /api/cours/specialite/{specialiteId}:
    get:
      summary: Lister les cours d'une spécialité
      tags: [Cours]
      operationId: getCoursBySpecialite
      parameters:
        - name: specialiteId
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
      responses:
        '200':
          description: Liste des cours
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Cours'
  /api/cours/option/{optionId}:
    get:
      summary: Lister les cours d'une option
      tags: [Cours]
      operationId: getCoursByOption
      parameters:
        - name: optionId
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
      responses:
        '200':
          description: Liste des cours
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Cours'

  # ========== DEVOIRS ==========
...